SamServ/requirements_2.txt
transformers==4.28.1
Pillow==9.5.0
timm==0.6.13
Flask==2.2.3
gunicorn==20.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cu118
git+https://github.com/facebookresearch/segment-anything.git